Refactor: rename Interfaces folder to the Abstractions folder at the Domain project
